refactor(chars): Update model/form custom Charfields #48
This commit contains a lot of small changes that are required to complete the next few issues. This includes commenting out the custom tag creation and automatic table updates due to circular references after this refactor.

Print statements have been left in for the completion of the next few issues

closes #48