feat(theme): Add dropdown theme functionality #119
Add setting to allow key theme and path value to choose a theme.
The theme is can be selected on a per field basis in the form init and
assign the theme name to the field widget.
Default tag_me theme is added if the dev does not create new themes.

closes #119