fix(init): Update method for passing extra attrs #135
If tag-me is used within a view that overrides the get_form_kwargs
and injects a 'user' key the following error is raised!
Form() got multiple values for keyword argument 'user'

The fix was to remove the direct passing of user to the init, add a check in
form kwargs and only add the user if it was not present.

closes #135