chore(conf): Add help and mgmt default settings #184
The help and management tools require urls that should be user
configurable. These are added to the bottom of the searchbar menu
so that if the dev desires, they can add custom urls. By default these
are empty so the section for display of the urls in the menu will be
hidden.