feat(ui): Improve ux and small screen appearance #184 (#186)
* docs(how): Add new and update existing #184

* feat(widget): Improve custom template handling #185

Renamed setting from theme to template to reflect its true purpose.
Make use of get_template rather than assigning the template to
self.template_name. Improves robustness.

closes #185

* chore(conf): Add help and mgmt default settings #184

The help and management tools require urls that should be user
configurable. These are added to the bottom of the searchbar menu
so that if the dev desires, they can add custom urls. By default these
are empty so the section for display of the urls in the menu will be
hidden.

* chore(context): Add help and mgmt extracted to context #184

closes #184

* docs(how-to): Add conf for help & mgmt urls #184

closes #184

* feat(ui): Improve ux and small screen appearance #184

closes #184