feat(templatetag):Added template tag to display pills #188
This commit adds a template tag to create pills from a tag field that are
consistent with the pills in the form.

I was unable to run tests to make sure that I have not broken anything, or
added tests for these template tags. When following the contributing
guidelines, I was unable to run the blog, or tests. See #227

closes #188